run show current_timestamp it case in standalone mode
